backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
Pillow==10.1.0
qrcode[pil]==7.4.2
loguru==0.7.2
